requirements.txt
# Worker app dependencies
pdfplumber==0.11.7
pdfminer.six==20250506
pypdfium2==4.30.0
mysql-connector-python==9.4.0
python-dotenv==1.1.1
structlog==25.4.0
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-distro==0.59b0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-semantic-conventions==0.59b0
grpcio==1.76.0
googleapis-common-protos==1.72.0
protobuf==6.33.1
requests==2.32.5

# WebAPI dependencies
fastapi==0.119.0
uvicorn==0.38.0
pydantic==2.12.3

# Testing
pytest
hypothesis
httpx
